refactor: introduce cache factory
