+++ alumnium/area.py
@@ -53,15 +53,15 @@ class Area:
         Raises:
             AssertionError: If the verification fails.
         """
-        result = self.client.retriever_agent.invoke(
+        explanation, value = self.client.retriever_agent.invoke(
             f"Is the following true or false - {statement}",
             self.accessibility_tree.to_xml(),
             title=self.driver.title,
             url=self.driver.url,
             screenshot=self.driver.screenshot if vision else None,
         )
-        assert result.value, result.explanation
-        return result.explanation
+        assert value, explanation
+        return explanation
 
     def get(self, data: str, vision: bool = False) -> Data:
         """
@@ -74,10 +74,11 @@ class Area:
         Returns:
             Data: The extracted data loosely typed to int, float, str, or list of them.
         """
-        return self.client.retriever_agent.invoke(
+        _, value = self.client.retriever_agent.invoke(
             data,
             self.accessibility_tree.to_xml(),
             title=self.driver.title,
             url=self.driver.url,
             screenshot=self.driver.screenshot if vision else None,
-        )[1]
+        )
+        return value
