+++ alumnium/cache/sqlite_cache.py
@@ -19,8 +19,8 @@ from sqlalchemy import (
 from sqlalchemy.orm import Session, declarative_base, relationship
 from xxhash import xxh3_128_hexdigest
 
-from .logutils import get_logger
-from .models import Model
+from ..logutils import get_logger
+from ..models import Model
 
 logger = get_logger(__name__)
 Base = declarative_base()
@@ -48,7 +48,7 @@ class CacheEntry(Base):
     __table_args__ = (UniqueConstraint("model_config_id", "hashed_request", name="uix_cache_entry"),)
 
 
-class Cache(BaseCache):
+class SQLiteCache(BaseCache):
     def __init__(self, db_path: str = ".alumnium-cache.sqlite"):
         self.engine = create_engine(f"sqlite:///{getcwd()}/{db_path}", connect_args={"timeout": 15})
         Base.metadata.create_all(self.engine)
